refactor: rename home-header folder to homeHeader
Use camelCase naming convention for folder consistency.

Fixes #402
