+++ app/src/components/homeHeader/HeaderContent.tsx
