+++ app/src/components/homeHeader/HeaderLogo.tsx
