+++ app/src/tests/fixtures/components/homeHeader/HeaderLogoMocks.tsx
