+++ app/src/tests/unit/components/homeHeader/HeaderLogo.test.tsx
@@ -1,7 +1,7 @@
 import { renderWithCountry, screen } from '@test-utils';
 import { beforeEach, describe, expect, test, vi } from 'vitest';
-import HeaderLogo from '@/components/home-header/HeaderLogo';
-import { TEST_COUNTRY_IDS } from '@/tests/fixtures/components/home-header/CountrySelectorMocks';
+import HeaderLogo from '@/components/homeHeader/HeaderLogo';
+import { TEST_COUNTRY_IDS } from '@/tests/fixtures/components/homeHeader/CountrySelectorMocks';
 
 const mockNavigate = vi.fn();
