+++ app/src/tests/unit/components/homeHeader/NavItem.test.tsx
@@ -1,10 +1,10 @@
 import { render, screen, userEvent } from '@test-utils';
 import { describe, expect, test, vi } from 'vitest';
-import NavItem from '@/components/home-header/NavItem';
+import NavItem from '@/components/homeHeader/NavItem';
 import {
   MOCK_DROPDOWN_NAV_ITEM,
   MOCK_SIMPLE_NAV_ITEM,
-} from '@/tests/fixtures/components/home-header/NavItemMocks';
+} from '@/tests/fixtures/components/homeHeader/NavItemMocks';
 
 describe('NavItem', () => {
   describe('Simple link behavior', () => {
